move relations style to grid
