+++ packages/client/src/pages/MainPage/containers/EntityDetailBox/EntityDetail/EntityDetailInverseRelations/EntityDetailInverseRelation/EntityDetailInverseRelation.tsx
