+++ packages/client/src/pages/MainPage/containers/EntityDetailBox/EntityDetail/EntityDetailInverseRelations/EntityDetailInverseRelations.tsx
