+++ packages/client/src/pages/MainPage/containers/EntityDetailBox/EntityDetail/EntityDetailInverseRelations/EntityDetailInverseRelationsStyles.tsx
