+++ packages/client/src/pages/MainPage/containers/EntityDetailBox/EntityDetail/EntityDetailRelations/EntityDetailRelationsStyles.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const StyledRelationsGrid = styled.div`
+  display: grid;
+  grid-template-columns: auto auto 1fr;
+  grid-gap: ${({ theme }) => theme.space[4]};
+`;
