--- packages/client/src/pages/MainPage/containers/EntityDetailBox/EntityDetail/EntityDetailRelationsSection/EntityDetailRelationsSectionStyles.tsx
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export const StyledRelationsWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
